refactor: compartmentalised heatmap feedback for more modular layout
